moved UML from MainWindow.cs to Data.cs
