+++ src/Patches.cs
@@ -7,62 +7,72 @@ using System.Xml.Linq;
 
 namespace cpatcher;
 
-public class CircloOPatch
+// everything what will be included in setup phase
+public readonly struct PatchInfo
+{
+    public string Name { get; init; } // aka internal name, init makes it so it cant be changes afterwards
+    public string DisplayName { get; init; }
+    public string Description { get; init; }
+    public string Version { get; init; }
+    public PatchInfo(string name, string displayName, string description, string version)
+    {
+        this.Name = name;
+        this.DisplayName = displayName;
+        this.Description = description;
+        this.Version = version;
+    }
+}
+
+public class Patch
 {
     public Guid Id { get; init; }
-    public string Name { get; init; } // aka internal name, init makes it so it cant be changes afterwards
-    public string DisplayName { get; set; }
-    public string Description { get; set; }
-    public Action Callback { get; set; }
-    public bool Required { get; set; }
+    public PatchInfo Info { get; init; }
+    public string CodePath { get; set; }
 
-    public CircloOPatch(string Name, string DisplayName, string Description, Action Callback, bool Required = false)
+    public Patch(string codePath, PatchInfo info)
     {
         this.Id = Guid.NewGuid();
-        this.Name = Name;
-        this.DisplayName = DisplayName;
-        this.Description = Description;
-        this.Callback = Callback;
-        this.Required = Required;
+        this.CodePath = codePath;
+        this.Info = info;
     }
 
     public void Execute()
     {
-        this.Callback(); // pass some additional arguments in like data and helper functions
+        // roslyn RunCSharpScript here
     }
 }
 
 public partial class MainWindow
 {
-    public List<CircloOPatch> Patches { get; set; } = new List<CircloOPatch>();
-    public List<CircloOPatch> SelectedPatches { get; set; } = new List<CircloOPatch>();
+    public List<Patch> Patches { get; set; } = new List<Patch>();
+    public List<Patch> SelectedPatches { get; set; } = new List<Patch>();
 
-    private CircloOPatch? FindByName(string name)
+    private Patch? FindByName(string name)
     {
         return Patches.Find(p => p.Name == name);
     }
 
-    public void NewPatch(CircloOPatch patch)
+    public void NewPatch(Patch patch)
     {
         Patches.Add(patch);
     }
 
     public void NewPatch(string name, string displayName, string description, bool required, Action callback)
     {
-        NewPatch(new CircloOPatch(name, displayName, description, callback, required));
+        NewPatch(new Patch(name, displayName, description, callback, required));
     }
 
     public void NewPatch(string name, bool required, Action callback)
     {
-        NewPatch(new CircloOPatch(name, name, name, callback, required));
+        NewPatch(new Patch(name, name, name, callback, required));
     }
 
     public void NewPatch(string name, Action callback)
     {
-        NewPatch(new CircloOPatch(name, name, name, callback, false));
+        NewPatch(new Patch(name, name, name, callback, false));
     }
 
-    public bool HasPatch(CircloOPatch patch)
+    public bool HasPatch(Patch patch)
     {
         return SelectedPatches.Any(p => p.Name == patch.Name) || FindByName(patch.Name).Required;
     }
@@ -74,7 +84,7 @@ public partial class MainWindow
 
     public void ExecuteAllPatches()
     {
-        foreach (CircloOPatch patch in Patches)
+        foreach (Patch patch in Patches)
         {
             if (HasPatch(patch))
             {
